fix: type for rootState
