+++ agoraStreaming/src/Navigation/Stack/index.tsx
@@ -4,9 +4,9 @@ import {
 } from '@react-navigation/stack';
 import React from 'react';
 
-import {Home} from '../Screens/Home';
-import {Live} from '../Screens/Live';
-import {RootStackParamList} from './types';
+import {Home} from '../../Screens/Home';
+import {Live} from '../../Screens/Live';
+import {HomeStackScreens, RootStackParamList} from './types';
 
 const Stack = createStackNavigator<RootStackParamList>();
 
@@ -16,7 +16,7 @@ export const MainStack = () => {
   return (
     <Stack.Navigator>
       <Stack.Screen
-        name="Home"
+        name={HomeStackScreens.Home}
         component={Home}
         options={options}
         // TODO: You need use this fragment after create user page or notification
@@ -28,7 +28,11 @@ export const MainStack = () => {
         //   },
         // }}
       />
-      <Stack.Screen name="Live" component={Live} options={options} />
+      <Stack.Screen
+        name={HomeStackScreens.Live}
+        component={Live}
+        options={options}
+      />
     </Stack.Navigator>
   );
 };
