+++ agoraStreaming/src/Navigation/Stack/types.ts
@@ -1,0 +1,21 @@
+import {Region} from 'react-native-maps';
+
+export enum HomeStackScreens {
+  Home = 'Home',
+  Live = 'Live',
+}
+
+export enum LiveType {
+  CREATE = 'create',
+  JOIN = 'join',
+}
+
+export type RootStackParamList = {
+  Home: undefined;
+  Live: {
+    type: LiveType;
+    channelId: string;
+    name?: string;
+    coords?: Region;
+  };
+};
