+++ agoraStreaming/src/Navigation/Tab/index.tsx
@@ -1,4 +1,4 @@
-import auth from '@react-native-firebase/auth';
+import auth, {FirebaseAuthTypes} from '@react-native-firebase/auth';
 import {GoogleSignin} from '@react-native-google-signin/google-signin';
 import {
   BottomTabNavigationOptions,
@@ -6,22 +6,21 @@ import {
 } from '@react-navigation/bottom-tabs';
 import {NavigationContainer} from '@react-navigation/native';
 import React, {useEffect} from 'react';
-import {Dimensions} from 'react-native';
 
 import {CalendarSvg} from '../../Icons/CalendarSvg';
 import {CircleSvg} from '../../Icons/CircleSvg';
 import {DiscoverSvg} from '../../Icons/DiscoverSvg';
 import {HomeSvg} from '../../Icons/HomeSvg';
 import {PlusSvg} from '../../Icons/PlusSvg';
-import {setUser} from '../../Redux/actions/AuthActions/AuthActions';
+import {setUser} from '../../Redux/actions/AuthActions';
 import {useAppDispatch, useAppSelector} from '../../Redux/hooks';
 import {selectUser} from '../../Redux/selectors/AuthSelectors';
 import {AuthScreen} from '../../Screens/Auth';
 import {ScreenCalendar} from '../../Screens/Calendar';
-import {MainStack} from '../index';
-import {ScreenOptionsType, TabNavigation, TabParamList} from '../types';
+import {MainStack} from '../Stack';
+import {styles} from './styles';
+import {ScreenOptionsType, TabNavigation, TabParamList} from './types';
 
-const {height} = Dimensions.get('window');
 const Tab = createBottomTabNavigator<TabParamList>();
 
 export const BottomTabs = () => {
@@ -46,29 +45,25 @@ export const BottomTabs = () => {
     },
     tabBarActiveTintColor: '#38a1e3',
     tabBarInactiveTintColor: '#fff',
-    tabBarStyle: {
-      borderTopColor: '#000',
-      height: height * 0.1,
-      backgroundColor: '#000',
-    },
+    tabBarStyle: styles.tabBar,
   });
 
   const dispatch = useAppDispatch();
   const user = useAppSelector(selectUser);
+
+  const onAuthStateChanged = (userInfo: FirebaseAuthTypes.User | null) => {
+    if (userInfo) {
+      const {uid, email, displayName} = userInfo;
+      dispatch(setUser({displayName, uid, email}));
+    }
+  };
 
   useEffect(() => {
     GoogleSignin.configure({
       webClientId:
         '1088468777432-7titji4f8tsu0oorpqfibu469sl8jvnj.apps.googleusercontent.com',
     });
-    auth().onAuthStateChanged((userInfo) => {
-      if (userInfo) {
-        const {uid, email, displayName} = userInfo;
-        dispatch(setUser({displayName, uid, email}));
-      } else {
-        console.log('User not logged in');
-      }
-    });
+    auth().onAuthStateChanged(onAuthStateChanged);
   }, []);
 
   if (!user) {
