+++ agoraStreaming/src/Navigation/Tab/types.ts
@@ -1,26 +1,5 @@
 import {BottomTabNavigationOptions} from '@react-navigation/bottom-tabs';
 import {ParamListBase, RouteProp} from '@react-navigation/native';
-import {Region} from 'react-native-maps';
-
-export enum LiveType {
-  CREATE = 'create',
-  JOIN = 'join',
-}
-
-export type RootStackParamList = {
-  Home: undefined;
-  Live: {
-    type: LiveType;
-    channelId: string;
-    name?: string;
-    coords?: Region;
-  };
-};
-
-export enum HomeStackScreens {
-  Home = 'Home',
-  Live = 'Live',
-}
 
 export enum TabNavigation {
   Main = 'Main',
