+++ agoraStreaming/src/Redux/store.ts
@@ -8,6 +8,6 @@ const rootReducer = combineReducers({
 
 export const store = createStore(rootReducer);
 
-export type RootState = ReturnType<typeof store.getState>;
+export type RootState = ReturnType<typeof rootReducer>;
 
 export type AppDispatch = typeof store.dispatch;
