--- agoraStreaming/src/Screens/Auth/types.ts
@@ -1,5 +1,0 @@
-import {Dispatch, SetStateAction} from 'react';
-
-export type AuthScreenPropsType = {
-  setInit: Dispatch<SetStateAction<boolean>>;
-};
