+++ agoraStreaming/src/Screens/Home/types.ts
@@ -2,7 +2,7 @@ import {RouteProp} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
 import {Region} from 'react-native-maps';
 
-import {HomeStackScreens, RootStackParamList} from '../../Navigation/types';
+import {HomeStackScreens, RootStackParamList} from '../../Navigation/Tab/types';
 
 export type HomeScreenProps = {
   navigation: StackNavigationPropNavigation;
