+++ agoraStreaming/src/Screens/Live/helpers/isBroadcaster.ts
@@ -1,4 +1,4 @@
-import {LiveType} from '../../../Navigation/types';
+import {LiveType} from '../../../Navigation/Tab/types';
 
 export const isBroadcasterFunction = (param: string) => {
   return param === LiveType.CREATE;
