+++ agoraStreaming/src/Screens/Live/types.ts
@@ -1,7 +1,7 @@
 import {RouteProp} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
 
-import {HomeStackScreens, RootStackParamList} from '../../Navigation/types';
+import {HomeStackScreens, RootStackParamList} from '../../Navigation/Tab/types';
 
 export type LiveScreenProps = {
   navigation: StackNavigationPropNavigation;
